test: almost able to load yaml optimizer
