+++ jina/optimizers/parameters.py
@@ -2,10 +2,10 @@ from typing import Optional, Sequence, Union
 
 import inspect
 
-from ..jaml import JAML
+from ..jaml import JAML, JAMLCompatibleSimple
 
 
-class OptimizationParameter:
+class OptimizationParameter(JAMLCompatibleSimple):
     def __init__(
         self,
         parameter_name: str,
@@ -18,33 +18,6 @@ class OptimizationParameter:
         else:
             self.env_var = env_var
         self.parameter_name = parameter_name
-
-    @classmethod
-    def to_yaml(cls, representer, data):
-        """Required by :mod:`pyyaml` """
-        tmp = data._dump_instance_to_yaml(data)
-        representer.sort_base_mapping_type_on_output = False
-        return representer.represent_mapping('!' + cls.__name__, tmp)
-
-    @staticmethod
-    def _dump_instance_to_yaml(instance):
-
-        attributes = inspect.getmembers(instance, lambda a: not (inspect.isroutine(a)))
-        return {
-            a[0]: a[1]
-            for a in attributes
-            if not (a[0].startswith('__') and a[0].endswith('__'))
-        }
-
-    @classmethod
-    def from_yaml(cls, constructor, node):
-        """Required by :mod:`pyyaml` """
-        return cls._get_instance_from_yaml(constructor, node)
-
-    @classmethod
-    def _get_instance_from_yaml(cls, constructor, node):
-        data = constructor.construct_mapping(node, deep=True)
-        return cls(**data)
 
 
 class IntegerParameter(OptimizationParameter):
@@ -76,9 +49,6 @@ class IntegerParameter(OptimizationParameter):
         }
 
 
-JAML.register(IntegerParameter)
-
-
 class FloatParameter(OptimizationParameter):
     def __init__(
         self,
@@ -106,9 +76,6 @@ class FloatParameter(OptimizationParameter):
         }
 
 
-JAML.register(FloatParameter)
-
-
 class UniformParameter(OptimizationParameter):
     def __init__(self, low: float, high: float, *args, **kwargs):
         super().__init__(*args, **kwargs)
@@ -122,9 +89,6 @@ class UniformParameter(OptimizationParameter):
             'low': self.low,
             'high': self.high,
         }
-
-
-JAML.register(UniformParameter)
 
 
 class LogUniformParameter(OptimizationParameter):
@@ -142,9 +106,6 @@ class LogUniformParameter(OptimizationParameter):
         }
 
 
-JAML.register(LogUniformParameter)
-
-
 class CategoricalParameter(OptimizationParameter):
     def __init__(
         self, choices: Sequence[Union[None, bool, int, float, str]], *args, **kwargs
@@ -155,9 +116,6 @@ class CategoricalParameter(OptimizationParameter):
 
     def to_optuna_args(self):
         return {'name': self.env_var, 'choices': self.choices}
-
-
-JAML.register(CategoricalParameter)
 
 
 class DiscreteUniformParameter(OptimizationParameter):
@@ -177,15 +135,7 @@ class DiscreteUniformParameter(OptimizationParameter):
         }
 
 
-JAML.register(DiscreteUniformParameter)
-
-
 def load_optimization_parameters(filepath: str):
-    JAML.register(IntegerParameter)
-    JAML.register(FloatParameter)
-    JAML.register(UniformParameter)
-    JAML.register(LogUniformParameter)
-    JAML.register(CategoricalParameter)
-    JAML.register(DiscreteUniformParameter)
+
     with open(filepath, encoding='utf8') as fp:
         return JAML.load(fp)
